[NuGet] Fix license expression links in License Acceptance dialog
Using the non-native XWT toolkit to display the dialog allows the
hyperlinks to be clicked. It also fixes the scrollbars showing a
whitebackground, even when not used.

Looked at adding the resize logic back, which would reduce the
height of the dialog based on the contents. However using the XWT
toolkit which is non-native (GTK#) the dialog is displayed first
and then resized so you can see it shrink. Instead the dialog does
not resize to fit its contents which leaves some empty space.